refacor: Add atomic design folder structure
