+++ src/components/Main/atoms/SampleDiv.js
@@ -1,0 +1,3 @@
+const SampleDiv = () => { };
+
+export default SampleDiv;
