+++ src/components/Main/atoms/SampleInnerText.js
@@ -1,0 +1,3 @@
+const SampleInnerText = () => { };
+
+export default SampleInnerText;
