+++ src/components/Main/atoms/SampleRefreshButton.js
@@ -1,0 +1,3 @@
+const SampleRefreshButton = () => { };
+
+export default SampleRefreshButton;
