+++ src/components/Main/atoms/SampleTypeText.js
@@ -1,0 +1,3 @@
+const SampleTypeText = () => { };
+
+export default SampleTypeText;
