+++ src/components/Main/molecules/SampleGroup.js
@@ -1,0 +1,3 @@
+const SampleGroup = () => { };
+
+export default SampleGroup;
