+++ src/components/Main/organisms/InstrumentSampleZone.js
@@ -1,0 +1,3 @@
+const InstrumentSampleZone = () => { };
+
+export default InstrumentSampleZone;
