+++ src/pages/EditSample/index.js
@@ -1,24 +1,19 @@
-import WaveSurfer from "wavesurfer.js";
-import TimelinePlugin from "wavesurfer.js/dist/plugin/wavesurfer.timeline";
-import RegionsPlugin from "wavesurfer.js/dist/plugin/wavesurfer.regions";
-import * as Tone from "tone";
+import { useState } from "react";
 import styled from "styled-components";
 
-import CreateWaveManager from "../../customWrapper/waveManager";
-import cutWaveUtil from "../../util/cutWave";
-import { waveSurferDefaultOptions } from "../../customWrapper/defaultOptions";
-import SampleWaveForm from "../../components/EditSample/SampleWaveForm";
+import SampleWaveFormContainer from "../../components/EditSample/SampleWaveFormContainer";
+import SampleWaveEditContainer from "../../components/EditSample/SampleWaveEditContainer";
 import Header from "../../components/common/Header";
 
 export default function EditSamplePage({ history }) {
+  const [waveHandler, setWaveHandler] = useState(null);
   const sampleUrl = history.location.state;
-  const waveManager =
-    CreateWaveManager(WaveSurfer, { TimelinePlugin, RegionsPlugin }, Tone, waveSurferDefaultOptions, { cutWaveUtil });
 
   return (
     <StyledEditSamplePage>
       <Header>/main</Header>
-      <SampleWaveForm waveManager={waveManager} sampleUrl={sampleUrl} />
+      <SampleWaveFormContainer waveHandler={waveHandler} setWaveHandler={setWaveHandler} sampleUrl={sampleUrl} />
+      <SampleWaveEditContainer waveHandler={waveHandler} />
     </StyledEditSamplePage>
   );
 }
